feat(handles): longer corner runs, clamped per side
corner_length 120 -> 160 in the user's config makes the corner zones span
38% of a side, at the edge bars' expense — but doing that surfaced a latent
hazard and a coupling, both fixed here.

The hazard: nothing capped the corner run against the window, so on a
window short enough for one side's two corner runs to meet, every point on
that side became a corner zone and its MIDPOINT resized diagonally.

The coupling: the obvious cap (against the window's shorter side) binds
both axes to the short one — measured on an 840x520 window, it silently
held the top edge's corner run at 28% when the config asked for 38%.

So the clamp is per SIDE, 0.45 of that side's own length, applied
identically in the shader's zone cut and the hit test's corner squares. A
long side carries the full configured run; a short one shortens; a 10%
edge bar always survives in the middle. Verified by gap-notch positions:
top edge at 38%/62% (the configured length), left edge at 45%/54% (its
cap), where the single cap had pinned both to 28%.

+++ src/server/window.rs
@@ -3815,7 +3815,11 @@ impl Window {
                 self.border.frame,
                 band_screen as f32,
                 (band_screen as f32 * layout.border_taper.clamp(0.0, 1.0)).max(2.0),
-                (px(cl) as f32).max(band_screen as f32),
+                // Floored at the band; the shader clamps the run per SIDE
+                // (0.45 of that side), so a long side carries the full
+                // configured length while a short one shortens. The hit test
+                // applies the same per-axis clamp.
+                (px(cl) as f64).max(band_screen) as f32,
                 px(g) as f32,
                 layout.border_swell_curve,
                 (layout.border_corner_bulge as f64).min(short_side * 0.3) as f32,
